Project invite serilog logging

+++ src/Shamyr.Expendity.Server.Service/Dtos/Expense/CreateExpenseDto.cs
@@ -5,7 +5,7 @@ namespace Shamyr.Expendity.Server.Service.Dtos.Expense
   public class CreateExpenseDto
   {
     public string Name { get; init; } = default!;
-    public double Value { get; init; }
+    public decimal Value { get; init; }
     public string? Description { get; init; }
     public DateTime? AddedUtc { get; init; }
     public int? TypeId { get; init; }
