+++ src/Shamyr.Expendity.Server.Service/Dtos/Summary/SummaryDto.cs
@@ -4,7 +4,7 @@ namespace Shamyr.Expendity.Server.Service.Dtos.Summary
 {
   public class SummaryDto
   {
-    public double FullSum { get; init; }
+    public decimal FullSum { get; init; }
     public ICollection<SummaryEntryDto> Entries { get; init; } = default!;
   }
 }
