+++ src/Shamyr.Expendity.Server.Service/Models/Expense/ExpenseModel.cs
@@ -6,7 +6,7 @@ namespace Shamyr.Expendity.Server.Service.Models.Expense
   {
     public int Id { get; init; }
     public string Name { get; init; } = default!;
-    public double Value { get; init; }
+    public decimal Value { get; init; }
     public string? Description { get; init; }
     public DateTime AddedUtc { get; init; }
     public int? TypeId { get; init; }
