+++ src/Shamyr.Expendity.Server.Service/Models/Summary/SummaryModel.cs
@@ -4,7 +4,7 @@ namespace Shamyr.Expendity.Server.Service.Models.Summary
 {
   public class SummaryModel
   {
-    public double FullSum { get; init; }
+    public decimal FullSum { get; init; }
     public ICollection<SummaryEntryModel> Entries { get; init; } = default!;
   }
 }
